pyproject.toml
[project]
name = "Crypto-Person"
version = "0.1.0"
description = "A crypto agent"
authors = [
    {name = "Dewa Adji"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain_core==0.3.72",
    "langchain_openai==0.3.27",
    "langchain_tavily==0.2.6",
    "langgraph==0.6.2",
    "langsmith==0.4.4",
    "pydantic==2.11.7",
    "python-dotenv==1.0.1",
    "requests==2.32.4",
    "sseclient==0.0.27",
    "sseclient_py==1.8.0",
    "sympy==1.14.0",
    "tavily-python>=0.5.0"
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-dir]
"src" = "src"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
